feedsender-app/requirements.txt
certifi==2020.6.20
cffi==1.14.1
click==7.1.2
cryptography==3.0
decorator==4.4.2
fastapi==0.60.1
h11==0.9.0
httptools==0.1.1
pycparser==2.20
pydantic==1.6.1
python-telegram-bot==12.8
six==1.15.0
starlette==0.13.6
tornado==6.0.4
uvicorn==0.11.6
uvloop==0.14.0
websockets==8.1
